LayoutTests/platform/wpe/js/dom/regress-157246-expected.txt

CONSOLE MESSAGE: line 2: SyntaxError: Unexpected token '|'. Expected an opening '(' before a function's parameter list.
JSRegress/regress-157246

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL caught exception: SyntaxError: Unexpected token '|'. Expected an opening '(' before a function's parameter list.
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
